requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
uv==0.2.18
